Remove connect and action
